Improve authorization semantics
