+++ README.md
@@ -4,5 +4,4 @@ WIP...
 
 ## Candidates for GEM extraction
 
-- Authorizer pattern
 - Searchable ActiveStorage metadata
