--- app/authorizers/users_authorizer.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class UsersAuthorizer < Perm::Authorized
-  alias_method :user, :subject
-
-  def can_admin_system?
-    user.has_role?(ENUMS::USER_ROLES::ADMINISTRATOR)
-  end
-end
