+++ app/models/concerns/authorizers/imageable.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Authorizers
+  module Imageable
+    def can_view_imageable?(imageable)
+      return true if can_admin_system?
+      user == imageable
+    end
+  end
+end
