+++ app/models/concerns/authorizers/images.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Authorizers
+  module Images
+    def can_update_image?(image)
+      return true if can_admin_system?
+      image.record == user
+    end
+
+    def can_destroy_image?(image)
+      return true if can_admin_system?
+      image.record == user
+    end
+  end
+end
